+++ wordpress-username-checker/templates/username_check_form.php
@@ -1,7 +1,0 @@
-<?php
-
-/**
- *
- */
-
-$username_field = sprintf( '<input type="text" name=')