+++ changelog.md
@@ -1,3 +1,7 @@
 # Changelog
 ##  Version 1.0.0 --  Initial Release
---This is the initial release of the plugin. At this point it simply checks an email to see if it is valid or not.
+--   # This is the initial release of the plugin. At this point it simply checks an email to see if it is valid or not.
+##  Version 1.1.0 --  Minor Update
+--  # Made the email check form into a template  
+--  # Updated the email check form function name to email_check_front_end_shortcode  
+--  + Added the username check form function and shortcode